+++ .enc.env
@@ -1,0 +1,12 @@
+ICAL_URL=ENC[AES256_GCM,data:k6LkYJIxzheiUZmflgvVp7vuAhhIhgdDoBzkPZO7SjEWH37pNyE+tBCisrUgR2wQrPB+xh0746KKFhpqloZQksiOKXg+MWAzlyx1/8Jp2rM=,iv:+EZWLOafpHWu7DLzazxSRGmKNl/gT8e/lkm78DkdXk0=,tag:RwZk0pRjsLBblTVHm9M6jw==,type:str]
+FEED_TITLE=ENC[AES256_GCM,data:cbK4cPTSOf8oMOi8jp4QTI8rKIFT1xeyCIxXXjtYTG/exBQ0QdZXPdCSAlS5w8rgZ1bYp9RRqqHa8LbQ1MmEBc+h0uz4XTuxxr62lZEIrht8zMbTAvzguV7gGjpzgAQAoqlnx1xlBFOuuic=,iv:SoavDDlzoaQ6o2vVMNPYee3jOAWMrVJGC+OZ7FQl/FI=,tag:rCMxIERYyUuVc/JE8lAHyQ==,type:str]
+FEED_DESCRIPTION=
+FEED_AUTHOR_NAME=ENC[AES256_GCM,data:xOgux8TrbjNnOaiM0j8=,iv:dpQgPzz9J/C7rl9IkmlALt7gTRiuRn2dSsYbVbRYgx4=,tag:J57uO6F4fVtNSOBAWpQv6g==,type:str]
+FEED_AUTHOR_EMAIL=
+MODE=ENC[AES256_GCM,data:Rl9+ndQ20hTl8Qw=,iv:5CCqVO5pebizZQHHLj4Gd5K4j23HANmiIbiEcPfXJWg=,tag:HsgnYNTfq43Cjhc6cfH80g==,type:str]
+sops_age__list_0__map_enc=-----BEGIN AGE ENCRYPTED FILE-----\nYWdlLWVuY3J5cHRpb24ub3JnL3YxCi0+IFgyNTUxOSBXYVFsb2w0SGVXWW5kY0NU\nUDNKOW9ramtKb0ZIUDNyTVhSNjJlWTB4eG5nCjJUZElQbFA2S3Zac3V6SUdhSDN6\nVG8rM0ptYitpVW41R0M1VGZPanNTRTAKLS0tIGx1d2pJUDg5Z3dYanRJeHZaNGkz\nYU1KSDVZd3ZxK21KWmcvOElpNDczMVUK8ecCoc8Pe0xc0AxZQJYUj3EAwvB9Of/A\nIcc6hKWNXE5RnR41Weu2ZIemvlp8Xu9opH2k8FMRR8HFlxfSQYJpvg==\n-----END AGE ENCRYPTED FILE-----\n
+sops_age__list_0__map_recipient=age1hae0mml9ktpyjpmkcfahpzhk2xuckyn2gnw8tr3a6y6yz8gv9qyqr65n9h
+sops_lastmodified=2024-10-13T09:05:24Z
+sops_mac=ENC[AES256_GCM,data:8LZ1N4c+eIYnBPLrn9a/BlKVVsHr33Sd1rAKV3V7tEuk1nPe4orQ1n34WSLBfJJfT7LPDZGsr5zPzpL3jBKmDwjbtcyA5wfMCLpSa8Q6HvnK+j+Wmo9+Ec0atahdJgICxhd2nV7hMK5OxpEttsapXFJ3Tas2LemGfg6msHOC4Cc=,iv:V5IDJA9Jm5Bgyga/akwk4uhoZ9FLK1/LxwDf1h1vHbc=,tag:tRlBcxgeQisCEpG21htv0Q==,type:str]
+sops_unencrypted_suffix=_unencrypted
+sops_version=3.9.0
