+++ .sops.yaml
@@ -1,0 +1,3 @@
+---
+creation_rules:
+  - age: age1hae0mml9ktpyjpmkcfahpzhk2xuckyn2gnw8tr3a6y6yz8gv9qyqr65n9h
