+++ README.md
@@ -1,3 +1,14 @@
 # iCal to RSS Feed
 
 Because I'm not going to check my calendar all the time for news.
+
+## ENV
+
+```bash
+ICAL_URL=
+FEED_TITLE=
+FEED_DESCRIPTION=
+FEED_AUTHOR_NAME=
+FEED_AUTHOR_EMAIL=
+MODE=DEVELOPMENT # or `PRODUCTION`
+```
