+++ utils.go
@@ -1,31 +1,16 @@
 package main
 
 import (
-	"io"
-	"net/http"
 	"os"
+
+	"github.com/rs/zerolog"
 )
 
-func downloadFile(filepath string, url string) (err error) {
-	// Create the file
-	out, err := os.Create(filepath)
-	if err != nil {
-		return err
+func initTempFolder(tempPath string, logger zerolog.Logger) {
+	if _, err := os.Stat(tempPath); os.IsNotExist(err) {
+		err = os.Mkdir(tempPath, 0755)
+		if err != nil {
+			logger.Fatal().Err(err).Msgf("Error creating %s directory", tempPath)
+		}
 	}
-	defer out.Close()
-
-	// Get the data
-	resp, err := http.Get(url)
-	if err != nil {
-		return err
-	}
-	defer resp.Body.Close()
-
-	// Writer the body to file
-	_, err = io.Copy(out, resp.Body)
-	if err != nil {
-		return err
-	}
-
-	return nil
 }
